+++ pkg/models/common_test.go
@@ -20,11 +20,11 @@ import (
 	"testing"
 
 	"github.com/jmoiron/sqlx"
-	_ "github.com/mattn/go-sqlite3"
+	_ "modernc.org/sqlite"
 )
 
 func TestDbLoad(t *testing.T) {
-	db, err := sqlx.Connect("sqlite3", ":memory:")
+	db, err := sqlx.Connect("sqlite", ":memory:")
 	if err != nil {
 		t.Fatalf("an error '%s' was not expected when opening a stub database connection", err)
 	}
@@ -45,5 +45,4 @@ func TestDbLoad(t *testing.T) {
 	if err == nil {
 		t.Errorf("did not fail to load SQL test data")
 	}
-
 }